Added global scene playlist with new controls for skipping scenes
I have added a new global playlist. You now set the master playlist in the "SceneLoader" object in the common static prefab. This has a list of all songs. you can change song order here easily, and this enables us to now have a "next" and "previous" song buttons. Skip through songs with either A/B on the VR controls, or the "." and "," keys on the keyboard (easy to remember because they are under the > and < symbols).